Renamed the step modules.
