+++ crates/bld_runner/src/pipeline/version2.rs
@@ -1,7 +1,7 @@
 use crate::artifacts::v2::Artifacts;
 use crate::external::v2::External;
 use crate::platform::v2::Platform;
-use crate::step::version2::BuildStep;
+use crate::step::v2::BuildStep;
 use crate::token_context::v2::PipelineContext;
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
