+++ crates/bld_runner/src/runner/version2.rs
@@ -27,7 +27,7 @@ use crate::{
     external::v2::External,
     pipeline::version2::Pipeline,
     platform::{builder::TargetPlatformBuilder, v2::Platform},
-    step::version2::{BuildStep, BuildStepExec},
+    step::v2::{BuildStep, BuildStepExec},
     token_context::v2::PipelineContextBuilder,
     RunnerBuilder,
 };
